Extract CrossTaskConstantChangesIncrementalCompilationIntegrationTest
